+++ Plugins/Tweaks/Tweaks.hpp
@@ -19,7 +19,6 @@ class StringToIntBaseToAuto;
 class DeadCreatureFiresOnAreaExit;
 class PreserveActionsOnDMPossess;
 class FixGreaterSanctuaryBug;
-class AsyncLogFlush;
 
 class Tweaks : public NWNXLib::Plugin
 {
@@ -42,7 +41,6 @@ private:
     std::unique_ptr<DeadCreatureFiresOnAreaExit> m_DeadCreatureFiresOnAreaExit;
     std::unique_ptr<PreserveActionsOnDMPossess> m_PreserveActionsOnDMPossess;
     std::unique_ptr<FixGreaterSanctuaryBug> m_FixGreaterSanctuaryBug;
-    std::unique_ptr<AsyncLogFlush> m_AsyncLogFlush;
 };
 
 }